requirements.txt
pycryptodome==3.10.1
tqdm==4.61.0
termcolor==1.1.0
PyQt5
